cambios reporte factura

+++ src/main/java/ec/com/hoteleraWeb/safari/control/dao/FacturaTODao.java
@@ -1,13 +1,11 @@
 package ec.com.hoteleraWeb.safari.control.dao;
 
-import java.math.BigDecimal;
-
 import org.springframework.beans.factory.annotation.Autowired;
 
 import ec.com.hoteleraWeb.safari.control.entityAux.FacturaTO;
 import ec.com.hoteleraWeb.safari.utils.dao.GenericDao;
 
-public interface FacturaTODao extends GenericDao<FacturaTO, BigDecimal> {
+public interface FacturaTODao extends GenericDao<FacturaTO, Integer> {
 
 	@Autowired
 	public FacturaTO obtenerTotalesFactura(Integer codigoHotel, Integer codigoReservacion);
